app/requirements.txt
Django==5.0.7
django-cors-headers==4.4.0
psycopg2==2.9.9
djangorestframework==3.15.2
django-spaghetti-and-meatballs==0.4.2

#Swagger
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1 
referencing==0.35.1
rpds-py==0.18.1
uritemplate==4.1.1

celery==5.4.0
django_celery_results==2.5.1
redis==5.0.7
azure-identity==1.16.1
